requirements.txt
PyOpenGL
PyOpenGL_accelerate
numpy>=1.14.2
scipy==1.2.1
scikit_learn==0.18.2
pandas>=0.23.3
c3d>=0.3.0
fastdtw==0.3.2
matplotlib>=2.0.2
pygame>=1.9.3
tornado>=4.5.2
requests>=2.18.4
Pillow>=5.0.0
httplib2>=0.10.3
pytest>=3.4.2
tarjan>=0.2.3.2
PySignal==1.1.1
transformations
